+++ crates/nostr/src/nips/nip26.rs
@@ -152,7 +152,7 @@ where
 {
     let unhashed_token = DelegationToken::new(delegatee_pk, conditions);
     let hashed_token = Sha256Hash::hash(unhashed_token.as_bytes());
-    let message = Message::from_slice(hashed_token.as_byte_array())?;
+    let message = Message::from_digest_slice(hashed_token.as_byte_array())?;
     Ok(delegator_keys.sign_schnorr_with_ctx(secp, &message, rng)?)
 }
 
@@ -186,7 +186,7 @@ where
 {
     let unhashed_token = DelegationToken::new(delegatee_public_key, conditions);
     let hashed_token = Sha256Hash::hash(unhashed_token.as_bytes());
-    let message = Message::from_slice(hashed_token.as_byte_array())?;
+    let message = Message::from_digest_slice(hashed_token.as_byte_array())?;
     secp.verify_schnorr(&signature, &message, &delegator_public_key)?;
     Ok(())
 }
@@ -724,7 +724,7 @@ mod tests {
         let unhashed_token: String =
             format!("nostr:delegation:{delegatee_public_key}:{conditions}");
         let hashed_token = Sha256Hash::hash(unhashed_token.as_bytes());
-        let message = Message::from_slice(hashed_token.as_byte_array()).unwrap();
+        let message = Message::from_digest_slice(hashed_token.as_byte_array()).unwrap();
 
         let verify_result =
             SECP256K1.verify_schnorr(&signature, &message, &delegator_keys.public_key());